* commit 'upstream/4.10.2':
  Imported Upstream version 4.10.2

Conflicts:
	c/src/lib/libbsp/sparc/shared/include/pci.h
	config.guess
	config.sub

+++ c/src/libchip/rtc/ds1375-rtc.h
@@ -1,23 +1,23 @@
 #ifndef DS1375_I2C_RTC_H
 #define DS1375_I2C_RTC_H
-/* $Id: ds1375-rtc.h,v 1.4 2008/09/07 03:38:37 ralf Exp $ */
+/* $Id: ds1375-rtc.h,v 1.6 2010/04/25 21:18:05 joel Exp $ */
 
 /* Driver for the Maxim 1375 i2c RTC (TOD only; very simple...) */
 
-/* 
+/*
  * Authorship
  * ----------
  * This software was created by
  *
  *     Till Straumann <strauman@slac.stanford.edu>, 2005-2007,
  * 	   Stanford Linear Accelerator Center, Stanford University.
- * 
+ *
  * Acknowledgement of sponsorship
  * ------------------------------
  * The software was produced by
  *     the Stanford Linear Accelerator Center, Stanford University,
  * 	   under Contract DE-AC03-76SFO0515 with the Department of Energy.
- * 
+ *
  * Government disclaimer of liability
  * ----------------------------------
  * Neither the United States nor the United States Department of Energy,
@@ -26,18 +26,18 @@
  * completeness, or usefulness of any data, apparatus, product, or process
  * disclosed, or represents that its use would not infringe privately owned
  * rights.
- * 
+ *
  * Stanford disclaimer of liability
  * --------------------------------
  * Stanford University makes no representations or warranties, express or
  * implied, nor assumes any liability for the use of this software.
- * 
+ *
  * Stanford disclaimer of copyright
  * --------------------------------
  * Stanford University, owner of the copyright, hereby disclaims its
  * copyright and all other rights in this software.  Hence, anyone may
- * freely use it for any purpose without restriction.  
- * 
+ * freely use it for any purpose without restriction.
+ *
  * Maintenance of notices
  * ----------------------
  * In the interest of clarity regarding the origin and status of this
@@ -46,9 +46,9 @@
  * or distributed by the recipient and are to be affixed to any copy of
  * software made or distributed by the recipient that contains a copy or
  * derivative of this software.
- * 
+ *
  * ------------------ SLAC Software Notices, Set 4 OTT.002a, 2004 FEB 03
- */ 
+ */
 
 #include <rtems.h>
 #include <libchip/rtc.h>
@@ -64,10 +64,10 @@ bool
 rtc_ds1375_device_probe( int minor );
 
 uint32_t
-rtc_ds1375_get_register( uint32_t port, uint8_t reg );
+rtc_ds1375_get_register( uintptr_t port, uint8_t reg );
 
 void
-rtc_ds1375_set_register( uint32_t port, uint8_t reg, uint32_t value );
+rtc_ds1375_set_register( uintptr_t port, uint8_t reg, uint32_t value );
 
 /*
  * BSP must supply string constant argument 'i2cname' which matches
@@ -84,11 +84,11 @@ rtc_ds1375_set_register( uint32_t port, uint8_t reg, uint32_t value );
 #define DS1375_RTC_TBL_ENTRY(i2cname) \
 {                                              	      \
 	sDeviceName:	"/dev/rtc",                       \
-	deviceType:		RTC_CUSTOM,                       \
-	pDeviceFns:		&rtc_ds1375_fns,                  \
+	deviceType:	RTC_CUSTOM,                       \
+	pDeviceFns:	&rtc_ds1375_fns,                  \
 	deviceProbe:	rtc_ds1375_device_probe,          \
-	ulCtrlPort1:	(uint32_t)(i2cname),              \
-	ulDataPort:		0,                                \
+	ulCtrlPort1:	(uintptr_t)(i2cname),             \
+	ulDataPort:	0,                                \
 	getRegister:	rtc_ds1375_get_register,          \
 	setRegister:	rtc_ds1375_set_register,          \
 }
